meetu/backend/requirements.txt
Django>=4.2
psycopg2-binary
django-allauth[socialaccount]
djangorestframework
gunicorn
django-environ
requests
PyJWT